CLOUD-17: Move observer to endpoints package
