+++ deployer/deployment-controller/src/test/java/fish/payara/cloud/deployer/endpoints/DeploymentOberverTest.java
@@ -1,8 +1,6 @@
 /*
- *  DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
- * 
- *  Copyright (c) [2020] Payara Foundation and/or its affiliates. All rights reserved.
- * 
+ * Copyright (c) 2020 Payara Foundation and/or its affiliates. All rights reserved.
+ *
  *  The contents of this file are subject to the terms of either the GNU
  *  General Public License Version 2 only ("GPL") or the Common Development
  *  and Distribution License("CDDL") (collectively, the "License").  You
@@ -11,23 +9,20 @@
  *  https://github.com/payara/Payara/blob/master/LICENSE.txt
  *  See the License for the specific
  *  language governing permissions and limitations under the License.
- * 
- *  When distributing the software, include this License Header Notice in each
- *  file and include the License.
- * 
+ *
  *  When distributing the software, include this License Header Notice in each
  *  file and include the License file at glassfish/legal/LICENSE.txt.
- * 
+ *
  *  GPL Classpath Exception:
  *  The Payara Foundation designates this particular file as subject to the "Classpath"
  *  exception as provided by the Payara Foundation in the GPL Version 2 section of the License
  *  file that accompanied this code.
- * 
+ *
  *  Modifications:
  *  If applicable, add the following below the License Header, with the fields
  *  enclosed by brackets [] replaced by your own identifying information:
  *  "Portions Copyright [year] [name of copyright owner]"
- * 
+ *
  *  Contributor(s):
  *  If you wish your version of this file to be governed by only the CDDL or
  *  only the GPL Version 2, indicate your decision by adding "[Contributor]
@@ -40,7 +35,7 @@
  *  only if the new code is made subject to such option by the copyright
  *  holder.
  */
-package fish.payara.cloud.deployer.process;
+package fish.payara.cloud.deployer.endpoints;
 
 import java.lang.reflect.Field;
 import java.lang.reflect.Type;
@@ -52,6 +47,13 @@ import javax.ws.rs.sse.OutboundSseEvent;
 import javax.ws.rs.sse.Sse;
 import javax.ws.rs.sse.SseBroadcaster;
 import javax.ws.rs.sse.SseEventSink;
+
+import fish.payara.cloud.deployer.endpoints.DeploymentObserver;
+import fish.payara.cloud.deployer.process.ChangeKind;
+import fish.payara.cloud.deployer.process.DeploymentProcessState;
+import fish.payara.cloud.deployer.process.Namespace;
+import fish.payara.cloud.deployer.process.ProcessAccessor;
+import fish.payara.cloud.deployer.process.StateChanged;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -64,13 +66,11 @@ public class DeploymentOberverTest {
     @Test
     public void obervationTest() throws IllegalAccessException, NoSuchFieldException {
         DeploymentObserver observer = new DeploymentObserver();
-        Field sseField = observer.getClass().getDeclaredField("sse");
-        sseField.setAccessible(true);
-        sseField.set(observer, new EventBuilder());
+        observer.sse = new EventBuilder();
         EventSink sink = new EventSink();
-        DeploymentProcessState state = new DeploymentProcessState(new Namespace("p", "s"), "foo", null);
+        DeploymentProcessState state = ProcessAccessor.createProcessWithName("foo");
         observer.addRequest(sink, state.getId());
-        StateChanged event = new StateChanged(state, ChangeKind.PROVISION_FINISHED);
+        StateChanged event = ProcessAccessor.makeEvent(state, ChangeKind.PROVISION_FINISHED);
         observer.eventListener(event);
         Assert.assertTrue(sink.isClosed());
